backend/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
html5lib==1.0.1
idna==2.8
psycopg2==2.8.1
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
urllib3==1.24.2
webencodings==0.5.1
